SeedingService implementacao - Popula a base de dados em Ambiente Desenv

+++ WebVendasMvc/WebVendasMvc/Models/Seller.cs
@@ -14,7 +14,7 @@ namespace WebVendasMvc.Models
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
         public Department Department  { get; set; } // Um vendedor pode esta em um departamento
-        public decimal BaseSalary { get; set; }
+        public double BaseSalary { get; set; }
         public ICollection<SallersRecord> Sales { get; set; } = new List<SallersRecord>();
 
 
@@ -22,7 +22,7 @@ namespace WebVendasMvc.Models
         {
 
         }
-        public Seller(int id, string name, string email, DateTime birthDate, decimal baseSalary, Department department)
+        public Seller(int id, string name, string email, DateTime birthDate, double baseSalary, Department department)
         {
             Id = id;
             Name = name;
